requirements.txt
streamlit
pandas
murf-api
plotly
